+++ src/Contract/JsonSchema/Describer/WithAttributeDescriber.php
@@ -14,6 +14,7 @@ namespace Symfony\AI\Platform\Contract\JsonSchema\Describer;
 use Symfony\AI\Platform\Contract\JsonSchema\Attribute\With;
 use Symfony\AI\Platform\Contract\JsonSchema\Factory;
 use Symfony\AI\Platform\Contract\JsonSchema\Subject\PropertySubject;
+use Symfony\AI\Platform\Exception\IOException;
 
 /**
  * @phpstan-import-type JsonSchema from Factory
@@ -23,7 +24,17 @@ final class WithAttributeDescriber implements PropertyDescriberInterface
     public function describeProperty(PropertySubject $subject, ?array &$schema): void
     {
         foreach ($subject->getAttributes(With::class) as $attribute) {
-            $schema = array_replace_recursive($schema ?? [], array_filter((array) $attribute, static fn ($value) => null !== $value));
+            if ($attribute->ref) {
+                try {
+                    $with = json_decode(file_get_contents($attribute->ref), true, flags: \JSON_THROW_ON_ERROR);
+                } catch (\JsonException $e) {
+                    throw new IOException(\sprintf('Failed to load the schema from "%s"', $attribute->ref), 0, $e);
+                }
+            } else {
+                $with = array_filter((array) $attribute, static fn ($value) => null !== $value);
+            }
+
+            $schema = array_replace_recursive($schema ?? [], $with);
         }
     }
 }